# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ms-CRM OpenApi Documentation",
        "description": "ms-CRM OpenApi description",
        "contact": {
            "email": "florian@alt.bzh"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1/",
            "description": "Full ms-CRM API endPoint documentation"
        }
    ],
    "paths": {
...
        "/user/{id}": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "get user by id",
                "description": "Afficher la fiche d'un utilisateur",
                "operationId": "184be49ea2b205cb0daf821b82028d26",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "return user object",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "no user found"
                    }
                }
            },
            "put": {
                "tags": [
                    "user"
                ],
                "summary": "Update user",
                "description": "Mise à jour de la fiche d'un user",
                "operationId": "db60f1c0ed783c7c7189ec0d72f8a0ae",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id's user to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "return user object updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "user"
                ],
                "summary": "Delete user",
                "description": "Archivage de la fiche d'un user",
                "operationId": "c3bbcfd30a359ab76a95a40b67b3d502",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id's user to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "return user object",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not deleted"
                    }
                }
            }
        },
        "/user/withCompany/{id}": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "List of all companies for one user",
                "description": "Affichage de toutes les compagnies reliées à ce user",
                "operationId": "1d04bbfd2fb0d61d553857ff43384223",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "return user object",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No user store"
                    }
                }
            }
        },
        "/user/list": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Get list of all users",
                "description": "Liste de toutes les fiches des users",
                "operationId": "ec3a8fd6634348990d464a2742c09f6e",
                "responses": {
                    "200": {
                        "description": "return user object",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No user's list found"
                    }
                }
            }
        },
        "/user": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "Store user",
                "description": "Création de la fiche d'un user",
                "operationId": "2b009f4102c0cc8c13cfab19f04206ee",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "return user object",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No user store"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UserStoreRequest": {
                "title": "UserStoreRequest",
                "description": "User request body data",
                "required": [
                    "firstname",
                    "lastname",
                    "email",
                    "civility",
                    "status_id",
                    "particulier"
                ],
                "properties": {
                    "firstname": {
                        "title": "firstname",
                        "description": "firstname of user",
                        "type": "string",
                        "example": "Benjamin"
                    },
                    "lastname": {
                        "title": "lastname",
                        "description": "lastname of user",
                        "type": "string",
                        "example": "Franklin"
                    },
                    "email": {
                        "title": "email",
                        "description": "email of user",
                        "type": "string",
                        "example": "BenjaminFranklin@electricity.us"
                    },
                    "civility": {
                        "title": "civility",
                        "description": "civility of user, refer to customfield's db",
                        "type": "integer",
                        "example": "1"
                    },
                    "birthdate": {
                        "title": "birthdate",
                        "description": "birthday of user",
                        "type": "string",
                        "example": "21/12/2010"
                    },
                    "phone": {
                        "title": "phone",
                        "description": "phone of user",
                        "type": "integer",
                        "example": "123456789"
                    },
                    "status_id": {
                        "title": "status_id",
                        "description": "status of user, refer to customfield's db",
                        "type": "integer",
                        "example": "7"
                    },
                    "particulier": {
                        "title": "particulier",
                        "description": "particulier or in company",
                        "type": "boolean",
                        "example": "0"
                    },
                    "is_admin": {
                        "title": "is_admin",
                        "description": "admin or not",
                        "type": "integer",
                        "example": "0"
                    },
                    "company_id": {
                        "title": "company_id",
                        "description": "id of company attached at this user, if it's already known",
                        "type": "integer",
                        "example": "9"
                    },
                    "role": {
                        "title": "role",
                        "description": "role of this user in his company, refer to customfield's db",
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "title": "name",
                        "description": "name of company attached to this user, if it's a new company",
                        "type": "string",
                        "example": "Welch Fargo Company"
                    },
                    "address": {
                        "title": "address",
                        "description": "address' company attached to this user, if it's a new company",
                        "type": "string",
                        "example": "23, rue de la liberté"
                    },
                    "zip": {
                        "title": "zip",
                        "description": "zip of city attached to this user, if it's a new company",
                        "type": "integer",
                        "example": "49100"
                    },
                    "city": {
                        "title": "city",
                        "description": "city's address of company attached to this user, if it's a new company",
                        "type": "string",
                        "example": "Angers"
                    }
                },
                "type": "object"
            },
            "UserUpdateRequest": {
                "title": "UserUpdateRequest",
                "description": "User update request body data",
                "properties": {
                    "firstname": {
                        "title": "firstname",
                        "description": "firstname of user",
                        "type": "string",
                        "example": "Benjamin"
                    },
                    "lastname": {
                        "title": "lastname",
                        "description": "lastname of user",
                        "type": "string",
                        "example": "Franklin"
                    },
                    "email": {
                        "title": "email",
                        "description": "email of user",
                        "type": "string",
                        "example": "BenjaminFranklin@electricity.us"
                    },
                    "civility": {
                        "title": "civility",
                        "description": "civility of user, refer to customfield's db",
                        "type": "integer",
                        "example": "1"
                    },
                    "birthdate": {
                        "title": "birthdate",
                        "description": "birthday of user",
                        "type": "string",
                        "example": "21/12/2010"
                    },
                    "phone": {
                        "title": "phone",
                        "description": "phone of user",
                        "type": "integer",
                        "example": "123456789"
                    },
                    "status_id": {
                        "title": "status_id",
                        "description": "status of user, refer to customfield's db",
                        "type": "integer",
                        "example": "7"
                    },
                    "is_admin": {
                        "title": "is_admin",
                        "description": "admin or not",
                        "type": "integer",
                        "example": "0"
                    },
                    "particulier": {
                        "title": "particulier",
                        "description": "particulier or in company",
                        "type": "boolean",
                        "example": "0"
                    },
                    "company_id": {
                        "title": "company_id",
                        "description": "id of company attached at this user, if it's already known",
                        "type": "integer",
                        "example": "9"
                    },
                    "role": {
                        "title": "role",
                        "description": "role of this user in his company, refer to customfield's db",
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "title": "name",
                        "description": "name of company attached to this user, if it's a new company",
                        "type": "string",
                        "example": "Welch Fargo Company"
                    },
                    "address": {
                        "title": "address",
                        "description": "address' company attached to this user, if it's a new company",
                        "type": "string",
                        "example": "23, rue de la liberté"
                    },
                    "zip": {
                        "title": "zip",
                        "description": "zip of city attached to this user, if it's a new company",
                        "type": "integer",
                        "example": "49100"
                    },
                    "city": {
                        "title": "city",
                        "description": "city's address of company attached to this user, if it's a new company",
                        "type": "string",
                        "example": "Angers"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "firstname": {
                        "title": "firstname",
                        "description": "firstname of user",
                        "type": "string",
                        "example": "Benjamin"
                    },
                    "lastname": {
                        "title": "lastname",
                        "description": "lastname of user",
                        "type": "string",
                        "example": "Franklin"
                    },
                    "email": {
                        "title": "email",
                        "description": "email of user",
                        "type": "string",
                        "example": "BenjaminFranklin@electricity.us"
                    },
                    "civility": {
                        "title": "civility",
                        "description": "civility of user, refer to customfield's db",
                        "type": "integer",
                        "example": "1"
                    },
                    "birthday": {
                        "title": "birthday",
                        "description": "birthday of user",
                        "type": "string",
                        "example": "21/12/2010"
                    },
                    "phone": {
                        "title": "phone",
                        "description": "phone of user",
                        "type": "integer",
                        "example": "123456789"
                    },
                    "status_id": {
                        "title": "status_id",
                        "description": "status of user, refer to customfield's db",
                        "type": "integer",
                        "example": "7"
                    },
                    "is_admin": {
                        "title": "is_admin",
                        "description": "admin or not",
                        "type": "integer",
                        "example": "0"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            },
...
        }
    },
...
    "security": [
        []
    ]
}